Alteracao de double para bigdecimal Ajustes nos endpoints de dashboard

+++ src/main/java/com/vanderlei/cfp/entities/ContaBancaria.java
@@ -9,6 +9,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
@@ -31,9 +32,9 @@ public class ContaBancaria implements Serializable {
 
   private String numeroContaBancaria;
 
-  private Double limiteContaBancaria;
+  private BigDecimal limiteContaBancaria;
 
-  private Double saldoContaBancaria;
+  private BigDecimal saldoContaBancaria;
 
   private Boolean vincularSaldoBancarioNoTotalReceita;
 
@@ -53,8 +54,8 @@ public class ContaBancaria implements Serializable {
       final String id,
       final String nome,
       final String numeroContaBancaria,
-      final Double limiteContaBancaria,
-      final Double saldoContaBancaria,
+      final BigDecimal limiteContaBancaria,
+      final BigDecimal saldoContaBancaria,
       final Boolean vincularSaldoBancarioNoTotalReceita,
       final Boolean atualizarSaldoBancarioNaBaixaTitulo,
       final Usuario usuario) {
@@ -92,19 +93,19 @@ public class ContaBancaria implements Serializable {
     this.numeroContaBancaria = numeroContaBancaria;
   }
 
-  public Double getLimiteContaBancaria() {
+  public BigDecimal getLimiteContaBancaria() {
     return limiteContaBancaria;
   }
 
-  public void setLimiteContaBancaria(Double limiteContaBancaria) {
+  public void setLimiteContaBancaria(BigDecimal limiteContaBancaria) {
     this.limiteContaBancaria = limiteContaBancaria;
   }
 
-  public Double getSaldoContaBancaria() {
+  public BigDecimal getSaldoContaBancaria() {
     return saldoContaBancaria;
   }
 
-  public void setSaldoContaBancaria(Double saldoContaBancaria) {
+  public void setSaldoContaBancaria(BigDecimal saldoContaBancaria) {
     this.saldoContaBancaria = saldoContaBancaria;
   }
